extract shared configuration logic
